server/requirements.txt
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.24.0
pydantic>=2.10.1,<3.0.0
mcp>=1.2.0